requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pdfplumber>=0.10.0
pypdf>=3.17.0
google-cloud-storage>=2.14.0
google-generativeai>=0.3.0
openai>=1.10.0
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
